chore: add Cargo.toml
This is the same as the upstream (under "vendor/tiktoken/") but with the
Python dependencies removed.

To distinguish from the upstream version, I'm appended version metadata
(ie. the info after the "+" in the `version` number) as described here:

- https://doc.rust-lang.org/cargo/reference/resolver.html

This is just the abbreviated commit hash of the upstream project.

Note that we're not actually publishing this crate but only using it
internally, so we don't have to worry about the following:

> SemVer has the concept of “version metadata” with a plus in the
> version, such as 1.0.0+21AF26D3. This metadata is usually ignored, and
> should not be used in a version requirement. You should never publish
> multiple versions that differ only in the metadata tag